requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
xlsxwriter>=3.1.0
